+++ test/gmock-gtest-all.cc
@@ -1421,7 +1421,7 @@ bool ParseNaturalNumber(const ::std::string& str, Integer* number) {
 
 # else
 
-  typedef unsigned long long BiggestConvertible;  // NOLINT
+  typedef uint64_t BiggestConvertible;  // NOLINT
   const BiggestConvertible parsed = strtoull(str.c_str(), &end, 10);
 
 # endif  // GTEST_OS_WINDOWS && !defined(__GNUC__)
