+++ test/gtest/gtest.h
@@ -2999,7 +2999,7 @@ class TypeWithSize<8> {
   typedef unsigned __int64 UInt;
 #else
   typedef long long Int;  // NOLINT
-  typedef unsigned long long UInt;  // NOLINT
+  typedef uint64_t UInt;  // NOLINT
 #endif  // GTEST_OS_WINDOWS
 };
